requirements.txt
python-telegram-bot==20.6
sqlalchemy==2.0.25
alembic==1.13.1
python-dotenv==1.0.1
httpx==0.27.0